+++ packages/dart/lib/src/base/parse_constants.dart
@@ -1,7 +1,7 @@
 part of flutter_parse_sdk;
 
 // Library
-const String keySdkVersion = '5.1.3';
+const String keySdkVersion = '6.4.0';
 const String keyLibraryName = 'Flutter Parse SDK';
 
 // End Points
@@ -44,6 +44,7 @@ const String keyFileClassname = 'ParseFile';
 // Headers
 const String keyHeaderSessionToken = 'X-Parse-Session-Token';
 const String keyHeaderRevocableSession = 'X-Parse-Revocable-Session';
+const String keyHeaderCloudContext = 'X-Parse-Cloud-Context';
 const String keyHeaderUserAgent = 'user-agent';
 const String keyHeaderApplicationId = 'X-Parse-Application-Id';
 const String keyHeaderContentType = 'content-type';
@@ -58,6 +59,8 @@ const String keyParamSessionToken = 'sessionToken';
 // Storage
 const String keyParseStoreBase = 'flutter_parse_sdk_';
 const String keyParseStoreUser = '${keyParseStoreBase}user';
+const String keyParseStoreObjects = '${keyParseStoreBase}objects';
+const String keyParseStoreDeletes = '${keyParseStoreBase}deletes';
 const String keyParseStoreInstallation = '${keyParseStoreBase}installation';
 
 // Installation
@@ -81,5 +84,6 @@ const String keyVarInstallationId = 'installationId';
 // Error
 const String keyError = 'error';
 const String keyCode = 'code';
+const String keyNetworkError = 'NetworkError';
 
 const bool parseIsWeb = identical(0, 0.0);